pyproject.toml
[tool.poetry]
name = "trello_clone"
version = "0.1.0"
description = ""
authors = ["Azamat Galimzhanov <azamat@galimzhanov.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^2.2"
djangorestframework = "^3.11.0"
rules = "^2.2"
django-fsm = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
pytest-django = "^3.8.0"
pytest-factoryboy = "^2.0.3"
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
